requirements.txt
streamlit>=1.24.0
streamlit-authenticator>=0.2.2
streamlit-folium>=0.15.0
folium>=0.14.0
pandas>=2.0.3
#pysqlite3
bcrypt>=4.0.1
requests>=2.31.0
PyYAML
python-dotenv>=1.0.0
extra-streamlit-components>=0.1.60
stripe>=5.4.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
oauthlib>=3.2.2
geopy>=2.3.0
Pillow>=10.0.0
python-jose>=3.3.0
passlib>=1.7.4

